weather_api/requirements.txt
alembic==1.5.3
asgiref==3.3.1
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
Django==3.1.5
djangorestframework==3.12.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
Werkzeug==1.0.1
